aligned co-tuned vis and holo
generated an array (one with idxs and one logical) that corresponds to trials where tuned ensembles were shot during their pref oris. also generated an array (with idxs) for holoStims that are iso to a vis and visStims that are iso to a holo stim, which when combined could index into things like respMat, mRespTS, etc.